requirements.txt
easydict
einops
flask
h5py
hydra-core
lightning
modelscope
ninja
numpy==1.26.4
open3d
opencv-python
omegaconf
pybullet
pyyaml
scipy
timm
torch
torchaudio
torchvision
transformers
tqdm
wandb
zarr
chumpy
ultralytics